+++ CHANGES.rst
@@ -7,6 +7,12 @@ bin/buildoutChangelog of lizard-area
 
 - Added lizard-security to Area model.
 
+- Added functionality to synchronize area objcten with remote wfs.
+
+- Created management command and celery task to run sinchrnization as
+  periodic task.
+
+
 
 0.1.3 (2011-12-09)
 ------------------